requirements.txt
Package                   Version              Editable project location
------------------------- -------------------- --------------------------
addict                    2.4.0
appdirs                   1.4.4
asttokens                 2.4.1
attrs                     23.2.0
backcall                  0.2.0
blinker                   1.7.0
Bottleneck                1.3.7
certifi                   2024.2.2
cffi                      1.14.6
chardet                   4.0.0
click                     8.1.7
cloudpickle               3.0.0
cmake                     3.29.3
colorama                  0.4.6
coloredlogs               15.0.1
comm                      0.2.1
ConfigArgParse            1.7
crcmod                    1.7
cycler                    0.12.1
Cython                    0.28.2
dash                      2.7.0
dash-core-components      2.0.0
dash-html-components      2.0.0
dash-table                5.0.0
dask                      2023.5.0
debugpy                   1.6.7
decorator                 5.1.1
einops                    0.6.1
executing                 2.0.1
fastjsonschema            2.19.1
Flask                     3.0.2
flatbuffers               24.3.25
fsspec                    2024.2.0
future                    1.0.0
humanfriendly             10.0
idna                      2.10
importlib-metadata        7.0.1
importlib_resources       6.1.2
ipdb                      0.11
ipykernel                 6.29.3
ipython                   8.12.2
iso8601                   2.1.0
itsdangerous              2.1.2
jedi                      0.19.1
Jinja2                    3.1.3
joblib                    1.3.2
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
jupyter_client            8.6.0
jupyter_core              5.5.0
kiwisolver                1.4.5
locket                    1.0.0
Mako                      1.3.5
Markdown                  3.5.2
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.4.3
matplotlib-inline         0.1.6
mdurl                     0.1.2
mmcls                     0.25.0
mmcv-full                 1.4.0                /home/lmx/Users/mmcv-1.4.0
mmengine                  0.10.3
mmsegmentation            0.30.0
model-index               0.1.11
mpmath                    1.3.0
nbformat                  5.7.0
nest_asyncio              1.6.0
networkx                  3.1
ninja                     1.11.1.1
numexpr                   2.8.4
numpy                     1.24.4
onnx                      1.16.1
onnxruntime-gpu           1.15.1
onnxsim                   0.4.36
open3d                    0.17.0
open3d-python             0.3.0.0
opencv-python             4.9.0.80
opendatalab               0.0.4
openmim                   0.3.9
ordered-set               4.1.0
oss2                      2.4.0
packaging                 23.2
pandas                    2.0.3
parso                     0.8.3
partd                     1.4.1
pexpect                   4.9.0
pickleshare               0.7.5
pillow                    10.2.0
pip                       24.2
pkgutil_resolve_name      1.3.10
platformdirs              4.2.0
plotly                    5.19.0
plyfile                   0.6
prettytable               3.10.0
progressbar               2.5
prompt-toolkit            3.0.42
protobuf                  3.20.3
psutil                    5.9.0
ptyprocess                0.7.0
pure-eval                 0.2.2
py-cpuinfo                9.0.0
pyassimp                  3.3
pybind11                  2.12.0
pycocotools               2.0.0
pycparser                 2.21
pycuda                    2024.1
Pygments                  2.17.2
PyOpenGL                  3.1.1a1
pyparsing                 3.1.1
PyQt5                     5.15.10
PyQt5_sip                 12.13.0
pyrealsense2              2.54.2.5684
pyserial                  3.5
python-dateutil           2.8.2
pytools                   2024.1.3
pytz                      2023.3.post1
PyWavelets                1.4.1
PyYAML                    6.0.1
pyzmq                     25.1.2
referencing               0.33.0
requests                  2.25.1
rich                      13.7.1
rpds-py                   0.18.0
scikit-build              0.11.1
scikit-image              0.14.2
scikit-learn              1.1.3
scipy                     1.10.1
seaborn                   0.13.2
serial                    0.0.97
setuptools                68.2.2
sip                       6.8.3
six                       1.16.0
stack-data                0.6.2
sympy                     1.12.1
tabulate                  0.9.0
tenacity                  8.2.3
tensorboardX              2.6
tensorrt                  8.5.2.2
termcolor                 2.4.0
thop                      0.1.1.post2209072238
threadpoolctl             3.3.0
tomli                     2.0.1
toolz                     0.12.1
torch                     1.11.0
torchvision               0.13.0
tornado                   6.4
tqdm                      4.66.4
traitlets                 5.14.1
typing_extensions         4.10.0
tzdata                    2023.3
ultralytics               8.2.26
urllib3                   1.26.18
wcwidth                   0.2.13
Werkzeug                  3.0.1
wheel                     0.42.0
yacs                      0.1.8
yapf                      0.40.2
zipp                      3.17.0
